Updated examples
Fixed a bug where libset wouldn't create parent directories if .local/share wasn't available.
