--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "livePreview.defaultPreviewPath": "/target/doc/libdmd/index.html"
-}
