--- examples/client/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "client"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-libset = { path = "../.."}
-anyhow = "1.0.57"
